Add support for podman backend
Adds support for using podman as the container backend instead of
docker. podman eliminates many of the security concerns with using
docker, since the container runs in rootless mode. This means that even
if a malicious program were to escape from the container, it would only
have the permissions of the calling user.