Rework user commands to add to container commands
Reworks the way that commands specified by the user are wrapped by
Pyrex. Instead of completely replacing the container commands, the user
commands override and extend the container command defintions, taking
precedence over them (i.e. if the user excludes a command it will be
excluded even if the container includes it, and if the user includes it,
it will be included even if the container excludes it).

In order for this to work, the user command shims must be individually
written to execute the full path to the command inside the container,
since we don't know if it will be in $PATH.